+++ libs/contracts/__init__.py
@@ -1,0 +1,29 @@
+from .events import (
+    AccountChangedEvent,
+    AccountSnapshot,
+    AppointmentChangedEvent,
+    AppointmentSnapshot,
+    HistoryChangedEvent,
+    HistorySnapshot,
+    HospitalChangedEvent,
+    HospitalSnapshot,
+    TimetableAppointmentSnapshot,
+    TimetableChangedEvent,
+    TimetableSnapshot,
+    dump_event_payload,
+)
+
+__all__ = [
+    "AccountChangedEvent",
+    "AccountSnapshot",
+    "AppointmentChangedEvent",
+    "AppointmentSnapshot",
+    "HistoryChangedEvent",
+    "HistorySnapshot",
+    "HospitalChangedEvent",
+    "HospitalSnapshot",
+    "TimetableAppointmentSnapshot",
+    "TimetableChangedEvent",
+    "TimetableSnapshot",
+    "dump_event_payload",
+]